+++ flutter_inappwebview_forge_ios/ios/flutter_inappwebview_forge_ios/Sources/flutter_inappwebview_forge_ios/InAppWebView/InAppWebViewSettings.swift
@@ -34,6 +34,7 @@ public class InAppWebViewSettings: ISettings<InAppWebView> {
     var incognito = false
     var cacheEnabled = true
     var containerId: String? = nil
+    var proxySettings: [String: Any?]? = nil
     var transparentBackground = false
     var disableVerticalScroll = false
     var disableHorizontalScroll = false